refactor: BookService を定義
